Moved modal function into FormatCard
