+++ client/src/components/RelatedItems/Items.module.css
@@ -46,7 +46,8 @@
   font-size: 20px;
   position: absolute;
   top: 10px;
-  color: var(--main2-color);
+  color: #fff;
+  filter: drop-shadow(0px 2px 10px var(--main2-color))
 }
 
 .items-info {
